Change run API to ensure stable libxray_main flags
Old API allowed user to provide a file URL to the configuration.
This, in turn, allowed the user to provide a unique file URL on each
invocation, which doesn't work. Xray-Core can parse the flags only once
and a second invocation with different file could cause errors. With the
new API we control the exact parameters to Xray-Core so this is not an
issue.